+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.5.6
 
 * fixed method for DrawingDimension.get_tolerances()
+* fixed method for DrawingDimension.get_clip().
 
 ## 0.5.5